feat: add and implement common interfaces to register and update an user

+++ src/users/users.controller.ts
@@ -6,12 +6,15 @@ import {
   Patch,
   Param,
   Delete,
+  BadRequestException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
 import { Auth } from 'src/auth/decorators/auth.decorator';
 import { Role } from 'src/common/enums/role.enum';
+import { UpdateUserDto } from 'src/common/interfaces/update-user.dto';
+import { RegisterDto } from 'src/common/interfaces/register.dto';
+import { ActiveUser } from 'src/common/decorators/active-user.decorator';
+import { IActiveUser } from 'src/common/interfaces/active-user.interface';
 
 @Auth(Role.ADMIN)
 @Controller('users')
@@ -19,8 +22,8 @@ export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.usersService.create(createUserDto);
+  create(@Body() registerDto: RegisterDto) {
+    return this.usersService.create(registerDto);
   }
 
   @Get()
@@ -39,7 +42,9 @@ export class UsersController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(id);
+  remove(@Param('id') id: string, @ActiveUser() user: IActiveUser) {
+    // verifico que el usuario no pueda eliminar su propia cuenta
+    if (id !== user.sub) return this.usersService.remove(id);
+    else throw new BadRequestException('You cannot delete the account');
   }
 }
